requirements.txt
streamlit
pinecone-client
google-generativeai